+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Index.java
@@ -58,16 +58,16 @@ public class Index implements Nameable, Serializable {
   }
 
   /**
-   * Compares two sequences to check for duplicates or near matches. For sequences of different lengths, the additional characters
-   * on the longer sequence are ignored - "AAAA" and "AAAACC" are considered duplicates. For every non-matching character, edit
-   * distance increases by 1
+   * Compares two sequences to check for duplicates or near matches. The minimum sequence length to be considered must be determined before
+   * this function is called, as sequences are truncated to the minimum sequence length. For every non-matching character, edit distance
+   * increases by 1.
    * 
    * @param sequence1
    * @param sequence2
+   * @param minLength
    * @return edit distance between the two sequences with 0 meaning they are duplicates, 1 meaning they differ by one character, etc.
    */
-  public static int checkEditDistance(String sequence1, String sequence2) {
-    int minLength = Math.min(sequence1.length(), sequence2.length());
+  public static int checkEditDistance(String sequence1, String sequence2, int minLength) {
     int editDistance = minLength;
     for (int i = 0; i < minLength; i++) {
       if (sequence1.charAt(i) == sequence2.charAt(i)) {